+++ tasks/sandboxed_code_auditor/__init__.py
@@ -1,0 +1,1 @@
+"""Sandboxed code auditor deterministic task."""